functions/requirements.txt
pdfplumber
google-cloud-storage>=2.16.0
